--- auth-connector/src/main/java/com/sequenceiq/cloudbreak/auth/security/internal/InternalReady.java
@@ -1,11 +1,0 @@
-package com.sequenceiq.cloudbreak.auth.security.internal;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface InternalReady {
-}
