+++ auth-internal-api/build.gradle
@@ -1,0 +1,15 @@
+buildscript {
+  repositories {
+    mavenLocal()
+    mavenCentral()
+    maven { url = "$repoUrl" }
+    jcenter()
+  }
+  dependencies {
+    classpath "org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}"
+  }
+}
+
+dependencies {
+  compile project(':common')
+}
