+++ auth-internal/build.gradle
@@ -1,0 +1,20 @@
+buildscript {
+  repositories {
+    mavenLocal()
+    mavenCentral()
+    maven { url = "$repoUrl" }
+    jcenter()
+  }
+  dependencies {
+    classpath "org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}"
+  }
+}
+
+dependencies {
+  compile project(':auth-internal-api')
+
+  compile             group: 'org.springframework.boot',  name: 'spring-boot-starter-jersey', version: '2.0.3.RELEASE'
+
+  testCompile         group: 'org.mockito',               name: 'mockito-core',               version: mockitoVersion
+  testImplementation  group: 'org.assertj',               name: 'assertj-core',               version: assertjVersion
+}
