+++ authorization-common-api/build.gradle
@@ -20,6 +20,7 @@ jar {
 }
 
 dependencies {
+  compile project(':auth-internal-api')
   compile group: 'org.slf4j',                     name: 'slf4j-api',                      version: slf4jApiVersion
   compile group: 'org.apache.commons',            name: 'commons-lang3',                  version: apacheCommonsLangVersion
   compile group: 'org.glassfish.jersey.core',     name: 'jersey-client',                  version: jerseyCoreVersion
