+++ authorization-common/src/test/java/com/sequenceiq/authorization/service/PermissionCheckServiceTest.java
@@ -29,8 +29,8 @@ import com.sequenceiq.authorization.annotation.CheckPermissionByAccount;
 import com.sequenceiq.authorization.annotation.DisableCheckPermissions;
 import com.sequenceiq.authorization.annotation.InternalOnly;
 import com.sequenceiq.authorization.resource.AuthorizationResourceAction;
+import com.sequenceiq.cloudbreak.auth.ReflectionUtil;
 import com.sequenceiq.cloudbreak.auth.ThreadBasedUserCrnProvider;
-import com.sequenceiq.cloudbreak.auth.security.internal.ReflectionUtil;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PermissionCheckServiceTest {
