+++ core/src/main/java/com/sequenceiq/cloudbreak/controller/v4/DatalakeV4Controller.java
@@ -12,7 +12,6 @@ import com.sequenceiq.authorization.annotation.InternalOnly;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.common.StackType;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.DatalakeV4Endpoint;
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.StackViewV4Responses;
-import com.sequenceiq.cloudbreak.auth.security.internal.InternalReady;
 import com.sequenceiq.cloudbreak.auth.security.internal.TenantAwareParam;
 import com.sequenceiq.cloudbreak.exception.BadRequestException;
 import com.sequenceiq.cloudbreak.service.workspace.WorkspaceService;
@@ -20,7 +19,6 @@ import com.sequenceiq.distrox.v1.distrox.StackOperations;
 
 @Controller
 @InternalOnly
-@InternalReady
 public class DatalakeV4Controller implements DatalakeV4Endpoint {
 
     @Lazy
