+++ core/src/main/java/com/sequenceiq/cloudbreak/controller/v4/StackV4Controller.java
@@ -39,7 +39,6 @@ import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.upgrade.Upgrade
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.upgrade.UpgradeV4Response;
 import com.sequenceiq.cloudbreak.auth.security.internal.AccountId;
 import com.sequenceiq.cloudbreak.auth.security.internal.InitiatorUserCrn;
-import com.sequenceiq.cloudbreak.auth.security.internal.InternalReady;
 import com.sequenceiq.cloudbreak.auth.security.internal.TenantAwareParam;
 import com.sequenceiq.cloudbreak.domain.stack.Stack;
 import com.sequenceiq.cloudbreak.exception.BadRequestException;
@@ -50,7 +49,6 @@ import com.sequenceiq.flow.api.model.FlowIdentifier;
 
 @Controller
 @WorkspaceEntityType(Stack.class)
-@InternalReady
 @AuthorizationResource
 public class StackV4Controller extends NotificationController implements StackV4Endpoint {
