+++ environment/src/main/java/com/sequenceiq/environment/credential/v1/AuditCredentialV1Controller.java
@@ -14,7 +14,6 @@ import com.sequenceiq.authorization.annotation.CheckPermissionByAccount;
 import com.sequenceiq.authorization.resource.AuthorizationResourceAction;
 import com.sequenceiq.cloudbreak.auth.ThreadBasedUserCrnProvider;
 import com.sequenceiq.cloudbreak.auth.security.internal.AccountId;
-import com.sequenceiq.cloudbreak.auth.security.internal.InternalReady;
 import com.sequenceiq.cloudbreak.auth.security.internal.TenantAwareParam;
 import com.sequenceiq.cloudbreak.cloud.response.CredentialPrerequisitesResponse;
 import com.sequenceiq.cloudbreak.event.ResourceEvent;
@@ -30,7 +29,6 @@ import com.sequenceiq.environment.credential.v1.converter.CredentialToCredential
 import com.sequenceiq.notification.NotificationController;
 
 @Controller
-@InternalReady
 @AuthorizationResource
 public class AuditCredentialV1Controller extends NotificationController implements AuditCredentialEndpoint {
