+++ environment/src/main/java/com/sequenceiq/environment/environment/v1/EnvironmentController.java
@@ -29,7 +29,6 @@ import com.sequenceiq.authorization.annotation.ResourceNameList;
 import com.sequenceiq.authorization.annotation.ResourceObject;
 import com.sequenceiq.authorization.resource.AuthorizationResourceAction;
 import com.sequenceiq.cloudbreak.auth.ThreadBasedUserCrnProvider;
-import com.sequenceiq.cloudbreak.auth.security.internal.InternalReady;
 import com.sequenceiq.cloudbreak.auth.security.internal.TenantAwareParam;
 import com.sequenceiq.cloudbreak.structuredevent.rest.annotation.AccountEntityType;
 import com.sequenceiq.common.api.telemetry.request.FeaturesRequest;
@@ -62,7 +61,6 @@ import com.sequenceiq.environment.environment.v1.converter.EnvironmentApiConvert
 import com.sequenceiq.environment.environment.v1.converter.EnvironmentResponseConverter;
 
 @Controller
-@InternalReady
 @Transactional(TxType.NEVER)
 @AuthorizationResource
 @AccountEntityType(Environment.class)
