+++ freeipa/src/main/java/com/sequenceiq/freeipa/controller/FreeIpaV1Controller.java
@@ -21,7 +21,6 @@ import com.sequenceiq.authorization.annotation.ResourceCrn;
 import com.sequenceiq.authorization.annotation.ResourceObject;
 import com.sequenceiq.authorization.resource.AuthorizationResourceAction;
 import com.sequenceiq.cloudbreak.auth.security.internal.AccountId;
-import com.sequenceiq.cloudbreak.auth.security.internal.InternalReady;
 import com.sequenceiq.cloudbreak.validation.ValidationResult;
 import com.sequenceiq.cloudbreak.validation.ValidationResult.State;
 import com.sequenceiq.freeipa.api.v1.freeipa.cleanup.CleanupRequest;
@@ -57,7 +56,6 @@ import com.sequenceiq.freeipa.util.CrnService;
 
 @Controller
 @AuthorizationResource
-@InternalReady
 public class FreeIpaV1Controller implements FreeIpaV1Endpoint {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(FreeIpaV1Controller.class);
