+++ freeipa/src/main/java/com/sequenceiq/freeipa/ldap/v1/LdapConfigV1Controller.java
@@ -14,7 +14,6 @@ import com.sequenceiq.authorization.annotation.CheckPermissionByResourceObject;
 import com.sequenceiq.authorization.annotation.ResourceCrn;
 import com.sequenceiq.authorization.annotation.ResourceObject;
 import com.sequenceiq.authorization.resource.AuthorizationResourceAction;
-import com.sequenceiq.cloudbreak.auth.security.internal.InternalReady;
 import com.sequenceiq.cloudbreak.auth.security.internal.TenantAwareParam;
 import com.sequenceiq.freeipa.api.v1.ldap.LdapConfigV1Endpoint;
 import com.sequenceiq.freeipa.api.v1.ldap.model.create.CreateLdapConfigRequest;
@@ -28,7 +27,6 @@ import com.sequenceiq.notification.NotificationController;
 
 @Controller
 @Transactional(TxType.NEVER)
-@InternalReady
 @AuthorizationResource
 public class LdapConfigV1Controller extends NotificationController implements LdapConfigV1Endpoint {
     @Inject
