+++ settings.gradle
@@ -63,3 +63,6 @@ include 'audit-connector'
 include 'structuredevent-service-legacy'
 include 'structuredevent-service-cdp'
 include 'structuredevent-api-cdp'
+include 'auth-internal'
+include 'auth-internal-api'
+
